Allow building the old dynarec on x64 with VisualStudio2013

+++ src/r4300/recomph.h
@@ -41,7 +41,7 @@ void passe2(precomp_instr *dest, int start, int end, precomp_block* block);
 void init_assembler(void *block_jumps_table, int block_jumps_number, void *block_riprel_table, int block_riprel_number);
 void free_assembler(void **block_jumps_table, int *block_jumps_number, void **block_riprel_table, int *block_riprel_number);
 
-void gencallinterp(unsigned long addr, int jump);
+void gencallinterp(uintptr_t addr, int jump);
 
 void genupdate_system(int type);
 void genbnel(void);
