+++ src/r4300/x86/gr4300.c
@@ -350,7 +350,7 @@ void gendebug(void)
 }
 #endif
 
-void gencallinterp(unsigned long addr, int jump)
+void gencallinterp(uintptr_t addr, int jump)
 {
    free_all_registers();
    simplify_access();
